+++ VTK/Widgets/vtkBezierContourLineInterpolator.cxx
@@ -88,7 +88,7 @@ int vtkBezierContourLineInterpolator::InterpolateLine( vtkRenderer *vtkNotUsed(r
   while ( stackCount )
     {
     //process last point on stack
-    int recursionLevel = controlPointsStack[13*(stackCount-1)];
+    int recursionLevel = static_cast<int>(controlPointsStack[13*(stackCount-1)]);
     
     p1 = controlPointsStack + 13*(stackCount-1)+1;
     p2 = controlPointsStack + 13*(stackCount-1)+4;
